+++ src/widgets/mod.rs
@@ -1,5 +1,5 @@
 // Pushrod Widgets Library
-// Box Widget
+// Core Widgets
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,3 +12,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+#[macro_use]
+
+/// This is a `TextWidget`, which draws text in a clipped area.
+pub mod text_widget;